fix(pty): avoid stale selection copy
