+++ src/renderer/lib/pty/pty-keybindings.ts
@@ -37,9 +37,9 @@ export function shouldHandleInterruptFromTerminal(event: KeyEventLike): boolean 
 export function shouldCopySelectionFromTerminal(
   event: KeyEventLike,
   isMacPlatform: boolean,
-  hasSelection: boolean
+  hasSelection: boolean,
+  hasRecentSelection = false
 ): boolean {
-  if (!hasSelection) return false;
   if (event.type !== 'keydown') return false;
   if (event.key.toLowerCase() !== 'c') return false;
 
@@ -49,7 +49,9 @@ export function shouldCopySelectionFromTerminal(
   const shift = event.shiftKey === true;
 
   // Ctrl+Shift+C should copy on all platforms
-  if (ctrl && shift && !meta && !alt) return true;
+  if (ctrl && shift && !meta && !alt) return hasSelection || hasRecentSelection;
+
+  if (!hasSelection) return false;
 
   // Platform-specific default copy shortcuts
   if (isMacPlatform) {
@@ -57,27 +59,6 @@ export function shouldCopySelectionFromTerminal(
   }
 
   return ctrl && !meta && !shift && !alt;
-}
-
-const OSC_52_CLIPBOARD_TARGET = 'c';
-
-export function decodeOsc52ClipboardData(data: string): string | null {
-  const separatorIndex = data.indexOf(';');
-  if (separatorIndex === -1) return null;
-
-  const target = data.slice(0, separatorIndex);
-  if (target !== '' && !target.includes(OSC_52_CLIPBOARD_TARGET)) return null;
-
-  const encoded = data.slice(separatorIndex + 1).replace(/\s/g, '');
-  if (!encoded || encoded === '?') return null;
-
-  try {
-    const binary = atob(encoded);
-    const bytes = Uint8Array.from(binary, (char) => char.charCodeAt(0));
-    return new TextDecoder().decode(bytes);
-  } catch {
-    return null;
-  }
 }
 
 /**
